+++ .env-example
@@ -1,2 +1,4 @@
 PORT = 3000
 DB_HOST = mongodb+srv://user:password@cluster0.ltim2.mongodb.net/db-contacts?retryWrites=true&w=majority
+SALT = 
+JWT_SECRET_KEY =
