+++ api/service/schemas/userSchema.js
@@ -36,27 +36,4 @@ userSchema.methods.validPassword = function (password) {
 
 const UserSchema = mongoose.model("user", userSchema);
 
-// userMethods
-const createUser = async (userData) => {
-  return await UserSchema.create(userData);
-};
-
-const findUserByEmail = async (query) => {
-  return await UserSchema.findOne(query);
-};
-
-const findUserById = async (id) => {
-  return await UserSchema.findById(id);
-};
-
-const updateToken = async (id, token) => {
-  return await UserSchema.updateOne({ _id: id }, { token });
-};
-
-module.exports = {
-  UserSchema,
-  createUser,
-  findUserByEmail,
-  findUserById,
-  updateToken,
-};
+module.exports = UserSchema;
