+++ bin/server.js
@@ -5,14 +5,14 @@ const mongoose = require("mongoose");
 
 require("dotenv").config();
 
-const contactsRouter = require("../routes/api/contacts");
+const contactsRouter = require("../contacts/contacts.routes");
 const usersRouter = require("../users/users.routes");
 
 const app = express();
 
 const formatsLogger = app.get("env") === "development" ? "dev" : "short";
 
-require("../routes/service/token");
+require("../api/service/token");
 
 const PORT = process.env.PORT || 3000;
 const uriDB = process.env.DB_HOST;
